+++ src/App.vue
@@ -31,9 +31,14 @@
 					icon="icon-files-dark"
 					exact />
 				<AppNavigationItem
+					:to="{name: 'requestFiles'}"
+					:title="t('libresign', 'Request')"
+					icon="icon-rename"
+					exact />
+				<AppNavigationItem
 					:to="{name: 'validation'}"
-					icon="icon-file"
-					:title="t('libresign', 'Validate')" />
+					:title="t('libresign', 'Validate')"
+					icon="icon-file" />
 			</template>
 			<template #footer>
 				<AppNavigationSettings :title="t('libresign', 'Settings')">