+++ CHANGELOG.md
@@ -23,7 +23,8 @@ Types of changes:
 - Added integration with Approval app. Thanks to @eneiluj
 - Endpoint to list LibreSign files
 - Endpoint to attach files to LibreSign profile
-- Validate by app libresign 
+- Request sign by LibreSign App
+- Validate by LibreSign App
 
 # Changed
 - Increment of coverage on backend code